# Conflicts:
#	Backend-Donaruma/Backend-Donaruma.csproj
#	Backend-Donaruma/Controllers/PagosController.cs
#	Backend-Donaruma/Program.cs
#	Backend-Donaruma/bin/Debug/net8.0/Backend-Donaruma.deps.json
#	Backend-Donaruma/bin/Debug/net8.0/DonarumaAPI_DTOs.dll
#	Backend-Donaruma/bin/Debug/net8.0/DonarumaAPI_DTOs.pdb
#	Backend-Donaruma/obj/Backend-Donaruma.csproj.nuget.dgspec.json
#	Backend-Donaruma/obj/Debug/net8.0/Backend-Donaruma.AssemblyInfo.cs
#	Backend-Donaruma/obj/Debug/net8.0/Backend-Donaruma.AssemblyInfoInputs.cache
#	Backend-Donaruma/obj/Debug/net8.0/Backend-Donaruma.GeneratedMSBuildEditorConfig.editorconfig
#	Backend-Donaruma/obj/Debug/net8.0/Backend-Donaruma.assets.cache
#	Backend-Donaruma/obj/Debug/net8.0/Backend-Donaruma.csproj.FileListAbsolute.txt
#	Backend-Donaruma/obj/Debug/net8.0/Backend-Donaruma.genruntimeconfig.cache
#	Backend-Donaruma/obj/Debug/net8.0/rpswa.dswa.cache.json
#	Backend-Donaruma/obj/project.assets.json
#	Backend-Donaruma/obj/project.nuget.cache
#	DonarumaAPI_DTOs/bin/Debug/net8.0/DonarumaAPI_DTOs.dll
#	DonarumaAPI_DTOs/bin/Debug/net8.0/DonarumaAPI_DTOs.pdb
#	DonarumaAPI_DTOs/obj/Debug/net8.0/DonarumaAPI_DTOs.AssemblyInfo.cs
#	DonarumaAPI_DTOs/obj/Debug/net8.0/DonarumaAPI_DTOs.AssemblyInfoInputs.cache
#	DonarumaAPI_DTOs/obj/Debug/net8.0/DonarumaAPI_DTOs.GeneratedMSBuildEditorConfig.editorconfig
#	DonarumaAPI_DTOs/obj/Debug/net8.0/DonarumaAPI_DTOs.assets.cache
#	DonarumaAPI_DTOs/obj/Debug/net8.0/DonarumaAPI_DTOs.csproj.CoreCompileInputs.cache
#	DonarumaAPI_DTOs/obj/Debug/net8.0/DonarumaAPI_DTOs.dll
#	DonarumaAPI_DTOs/obj/Debug/net8.0/DonarumaAPI_DTOs.pdb
#	DonarumaAPI_DTOs/obj/Debug/net8.0/ref/DonarumaAPI_DTOs.dll
#	DonarumaAPI_DTOs/obj/Debug/net8.0/refint/DonarumaAPI_DTOs.dll
#	DonarumaAPI_DTOs/obj/DonarumaAPI_DTOs.csproj.nuget.dgspec.json
#	DonarumaAPI_DTOs/obj/project.assets.json
#	DonarumaAPI_DTOs/obj/project.nuget.cache
#	DonarumaAPI_Data/bin/Debug/net8.0/DonarumaAPI_DTOs.dll
#	DonarumaAPI_Data/bin/Debug/net8.0/DonarumaAPI_DTOs.pdb
#	DonarumaAPI_Data/bin/Debug/net8.0/DonarumaAPI_Data.dll
#	DonarumaAPI_Data/bin/Debug/net8.0/DonarumaAPI_Data.pdb
#	DonarumaAPI_Data/obj/Debug/net8.0/DonarumaAPI_Data.AssemblyInfo.cs
#	DonarumaAPI_Data/obj/Debug/net8.0/DonarumaAPI_Data.AssemblyInfoInputs.cache
#	DonarumaAPI_Data/obj/Debug/net8.0/DonarumaAPI_Data.GeneratedMSBuildEditorConfig.editorconfig
#	DonarumaAPI_Data/obj/Debug/net8.0/DonarumaAPI_Data.assets.cache
#	DonarumaAPI_Data/obj/Debug/net8.0/DonarumaAPI_Data.csproj.CoreCompileInputs.cache
#	DonarumaAPI_Data/obj/Debug/net8.0/DonarumaAPI_Data.dll
#	DonarumaAPI_Data/obj/Debug/net8.0/DonarumaAPI_Data.pdb
#	DonarumaAPI_Data/obj/Debug/net8.0/ref/DonarumaAPI_Data.dll
#	DonarumaAPI_Data/obj/Debug/net8.0/refint/DonarumaAPI_Data.dll
#	DonarumaAPI_Data/obj/DonarumaAPI_Data.csproj.nuget.dgspec.json
#	DonarumaAPI_Data/obj/project.assets.json
#	DonarumaAPI_Data/obj/project.nuget.cache
#	DonarumaAPI_Model/obj/Debug/net8.0/DonarumaAPI_Model.AssemblyInfo.cs
#	DonarumaAPI_Model/obj/Debug/net8.0/DonarumaAPI_Model.AssemblyInfoInputs.cache
#	DonarumaAPI_Model/obj/Debug/net8.0/DonarumaAPI_Model.GeneratedMSBuildEditorConfig.editorconfig
#	DonarumaAPI_Model/obj/Debug/net8.0/DonarumaAPI_Model.assets.cache
#	DonarumaAPI_Model/obj/DonarumaAPI_Model.csproj.nuget.dgspec.json
#	DonarumaAPI_Model/obj/project.assets.json
#	DonarumaAPI_Model/obj/project.nuget.cache

+++ Backend-Donaruma/Controllers/PagosController.cs
@@ -12,11 +12,11 @@ using System.Threading.Tasks;
 
 namespace Backend_Donaruma.Controllers
 {
-    // ─── 1. LOS MOLDES PARA ENTENDER A ANGULAR ───
     public class ItemCarrito
     {
         public string nombre { get; set; }
-        public long precio { get; set; } // Precio en pesos normales (ej. 500)
+        // 🔥 CAMBIO 1: Ahora aceptamos decimales (ej. 732.89)
+        public decimal precio { get; set; }
         public int cantidad { get; set; }
     }
 
@@ -71,7 +71,6 @@ namespace Backend_Donaruma.Controllers
             }
         }
 
-        // ─── 3. EL WEBHOOK SE QUEDA EXACTAMENTE IGUAL ───
         [HttpPost("webhook")]
         public async Task<IActionResult> StripeWebhook()
         {
